core: runtime capability contract + replay/live backends (P3-A)
RuntimeEngine is now the portable capability surface handlers program
against; no HANDLE or Win32 CONTEXT leaks into it:

- ThreadContextSnapshot: portable x64 rip/rsp/rbp/rax/eflags snapshot;
  the live adapter converts to/from the backend's native CONTEXT and
  set_thread_context is read-modify-write so untouched registers
  (including debug registers) survive.
- New required capabilities on RuntimeEngine: read_memory / write_memory
  (u64 addresses), get/set_thread_context, set/clear_hardware_breakpoint
  (DR0..DR3 slots with fail-closed validation: out-of-range slot, zero
  address, duplicate arm, clear-unarmed), pending_thread_id, and
  continue_thread which rejects thread-id mismatches and keeps the event
  pending on failure (exactly-once continue preserved).
- CapabilityRecord/CapabilityOp log: every capability operation records
  its monotonic sequence and thread id, so replay/live parity is asserted
  offline.
- ReplayRuntimeEngine: scripted contexts + breakpoint slots, unmapped
  reads stay fail-closed; inherent usize read_memory/write_memory removed
  in favor of the trait's u64 surface (themida test call site updated).
- DebuggerCoreEngine: capability impls over DebuggerCore with context
  conversion and breakpoint registers via context RMW.
- 7 new tests: record sequence/thread, unmapped read, unseeded context,
  slot validation, thread-id mismatch (both engines, pending retained),
  live context RMW preservation, and op-for-op capability-log parity
  between replay and live adapters.

Workspace: 32 suites green, 0 warnings; GTO feature check green;
cargo-deny offline check green.

+++ crates/packers/themida/src/plugin.rs
@@ -623,7 +623,7 @@ mod tests {
                 } => {
                     phases.push("guard_av");
                     let mut sample = [0u8; 4];
-                    eng.read_memory(*address as usize, &mut sample).unwrap();
+                    eng.read_memory(*address, &mut sample).unwrap();
                     assert_eq!(sample, [0xcc; 4]);
                     packer.note_guard_installed(&mut ctx);
                 }
